% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{activity_diag}
\alias{activity_diag}
\title{Visual diagnostic for activity data using MAVEn with baseline}
\usage{
activity_diag(
  maven_raw,
  metabolism_summary_cycle,
  activity_summary_cycle,
  interval = 120,
  outdir = "output",
  maven_experiment = "",
  out_filename = "ActivityDiagnostic",
  out_filetype = ".png"
)
}
\arguments{
\item{maven_raw}{Complete MAVEn dataset including baseline readings.}

\item{metabolism_summary_cycle}{Dataframe calculated by
\code{summarize_metabolism} with the "by_cycle" type indicated.}

\item{activity_summary_cycle}{Dataframe calculated by
\code{summarize_activity} with the "by_cycle" type indicated.}

\item{interval}{Time interval for analysis.}

\item{outdir}{Folder where figure should be stored.}

\item{maven_experiment}{Experiment name. Applied to both figures and files.}

\item{out_filename}{Figure name.}

\item{out_filetype}{Figure file type.}
}
\value{

}
\description{
Visual diagnostic for activity data using MAVEn with baseline
}
\examples{
activity_diag(maven_raw, metabolism_summary_cycle, activity_summary_cycle,
maven_experiment = "maven_example",
outdir = "output", out_filename = "ActivityDiagnostic", 
out_filetype = ".png")
}
