% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{activity_trend}
\alias{activity_trend}
\title{Plot animal activity trends.}
\usage{
activity_trend(
  animal_activity,
  maven_experiment = "",
  outdir = "output",
  out_filename = "ActivityTrends",
  out_filetype = ".png"
)
}
\arguments{
\item{animal_activity}{animal activity dataframe extracted from the MAVEn without
baseline using \code{extract_activity}.}

\item{maven_experiment}{Experiment name. Applied to both figures and files.}

\item{outdir}{Folder where figure should be stored.}

\item{out_filename}{Figure name.}

\item{out_filetype}{Figure file type.}
}
\description{
\code{activity_trend} graphcially displays the animal activity data by chamber.
The data are adjusted to a standardized measurement reading for illustration.
}
\examples{
#activity_trend(animal_activity, maven_experiment = "maven.example1")
}
