% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{co2_convertion}
\alias{co2_convertion}
\title{Convert CO2 MAVEn measurement to CO2 (ul/h)}
\usage{
co2_convertion(x)
}
\arguments{
\item{x}{list of CO2 values from MAVEn dataset}
}
\value{
list of converted CO2 measurement from ml/min to ul/hr.
}
\description{
Convert CO2 MAVEn measurement to CO2 (ul/h)
}
