% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{co2_convertion_median}
\alias{co2_convertion_median}
\title{Calculate median CO2(ul/h)}
\usage{
co2_convertion_median(x)
}
\arguments{
\item{x}{list of CO2 values from MAVEn dataset}
}
\value{

}
\description{
Calculate median CO2(ul/h)
}
