% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_maven}
\alias{evaluate_maven}
\title{Evaluate entire MAVEn dataset}
\usage{
evaluate_maven(datadir = "", outdir = NULL,
  maven_datafile = "./maven_output.csv", maven_experiment = "",
  interval = 60, activity_baseline = 0.01, activity_threshold = 1,
  figures = c("trend", "diag", "overview"))
}
\arguments{
\item{datadir}{data directory}

\item{outdir}{name of output directory. Defaults to "output"}

\item{maven_datafile}{name of data file to be evaluated. Must be ".csv".}

\item{maven_experiment}{name of experiment}

\item{interval}{specified interval for activity calculations. Must be in secoonds and defaults to 60.}

\item{activity_baseline}{Activity measurement baseline. Used to exclude values from dataset.}

\item{activity_threshold}{Activity threshold used for calculating activity status (e.g. inactive vs. active)}

\item{figures}{figures to be produced. Select from "overview", "diag", or "trend". Supply all three in list if you want all to be created.}
}
\description{
Evaluate entire MAVEn dataset
}
