% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_activity}
\alias{extract_activity}
\title{Extract animal activity data from MAVEn without baseline}
\usage{
extract_activity(
  maven.cycle,
  metabolism_summary_cycle,
  interval = "",
  activity_baseline = ""
)
}
\arguments{
\item{maven.cycle}{MAVEn dataset with cycles assigned. Must apply
\code{assign_cyclenumbers} to MAVEn dataset without baseline data.}

\item{metabolism_summary_cycle}{Summary dataset created by
\code{summarize_metabolism}.}

\item{interval}{Measurement interval for activity evaluation. Must be in
seconds. Recommend value less than 60 to stay within the instrument
metabolism measurement.}

\item{activity_baseline}{Baseline value for activity.}
}
\value{
Extracted animal activity dataset with median time and activity
  start/end times.
}
\description{
\code{extract_activity} extracts animal activity data from the cycle integrated
MAVEn dataset and calculated animal metabolism. It applies a measurement
adjustment to the time series for visualization with \code{activity_trend}.
}
\examples{
extract_activity(maven.cycle, metabolism_summary_cycle, 
interval = 60, activity_baseline = 0.01)
}
