% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_metabolism}
\alias{extract_metabolism}
\title{Extract animal metabolism data from MAVEn without baseline}
\usage{
extract_metabolism(maven.cycle)
}
\arguments{
\item{maven.cycle}{MAVEn dataset with cycles assigned. Must apply
\code{assign_cyclenumbers} to MAVEn dataset without baseline data.}
}
\value{
Extracted animal metabolism data.
}
\description{
\code{extract_metabolism} extracts animal metabolism data from the cycle
integrated MAVEn dataset. It applies a measurement adjustment to the time
series for visualization with \code{metabolism_trend}.
}
\examples{
extract_metabolism(maven)
}
