% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{maven_datatable}
\alias{maven_datatable}
\title{Create final MAVEn datatable for an experimental run}
\usage{
maven_datatable(
  metabolism_summary_cycle,
  activity_summary_cycle,
  outdir = NULL,
  out_filename = "ExperimentSummaryTable",
  maven_experiment = ""
)
}
\arguments{
\item{metabolism_summary_cycle}{Dataframe calculated by
\code{summarize_metabolism} with the "by_cycle" type indicated.}

\item{activity_summary_cycle}{Dataframe calculated by
\code{summarize_activity} with the "by_cycle" type indicated.}

\item{outdir}{Output directory. Defaults to 'output'}

\item{out_filename}{File name. Defaults to "ExperimentSummaryTable"}

\item{maven_experiment}{MAVEn experiment name.}
}
\value{

}
\description{
Create final MAVEn datatable for an experimental run
}
\examples{
#maven_datatable(metabolism_summary_cycle, activity_summary_cycle, 
#outdir = "output", out_filename = "ExperimentSummaryTable", 
#maven_experiment = "experiment")
}
