% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{metablism_trend}
\alias{metablism_trend}
\title{Plot fly metabolism trends.}
\usage{
metablism_trend(
  fly_metabolism,
  outdir = "output",
  out_filename = "MetabolismTrends",
  out_filetype = ".png"
)
}
\arguments{
\item{fly_metabolism}{Fly metabolism dataframe extracted from the MAVEn without baseline using \code{extract_metabolism}.}

\item{outdir}{Folder where figure should be stored.}

\item{out_filename}{Figure name.}

\item{out_filetype}{Figure file type.}
}
\value{

}
\description{
\code{metabolism_trend} graphcially displays the fly metabolism data by chamber. The data are adjusted to a standardized measurement reading for illustration.
}
\examples{
metablism_trend(fly_metabolism, outdir = "output", 
out_filename = "", out_filetype = ".png")
}
