% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{metabolism_diag}
\alias{metabolism_diag}
\title{Visual diagnostic for metabolism data using MAVEn with baseline}
\usage{
metabolism_diag(
  maven_raw,
  metabolism_summary_cycle,
  outdir = "output",
  out_filename = "MetabolismDiagnostic",
  out_filetype = ".png"
)
}
\arguments{
\item{maven_raw}{Complete MAVEn dataset including baseline readings.}

\item{metabolism_summary_cycle}{Dataframe calculated by \code{summarize_metabolism} with the "by_cycle" type indicated.}

\item{outdir}{Folder where figure should be stored.}

\item{out_filename}{Figure name.}

\item{out_filetype}{Figure file type.}
}
\value{

}
\description{
Visual diagnostic for metabolism data using MAVEn with baseline
}
