% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{metabolism_trend}
\alias{metabolism_trend}
\title{Plot animal metabolism trends.}
\usage{
metabolism_trend(
  animal_metabolism,
  maven_experiment = "",
  outdir = NULL,
  out_filename = "MetabolismTrends",
  out_filetype = ".png"
)
}
\arguments{
\item{animal_metabolism}{animal metabolism dataframe extracted from the MAVEn
without baseline using \code{extract_metabolism}.}

\item{maven_experiment}{Experiment name. Applied to both figures and files.}

\item{outdir}{Folder where figure should be stored.}

\item{out_filename}{Figure name.}

\item{out_filetype}{Figure file type.}
}
\value{

}
\description{
\code{metabolism_trend} graphcially displays the animal metabolism data by
chamber. The data are adjusted to a standardized measurement reading for
illustration.
}
\examples{
#metabolism_trend(animal_metabolism, outdir = "output",
#out_filename = "", out_filetype = ".png")
}
