% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_maven}
\alias{read_maven}
\title{Read MAVEn output into R.}
\usage{
read_maven(datadir = "data", maven_datafile = "maven_output.csv", baseline = F)
}
\arguments{
\item{datadir}{Directory where data is stored}

\item{maven_datafile}{MAVEn file in .csv format}

\item{baseline}{Toggle option to include baseline data. Default is `FALSE`}
}
\value{
MAVEn dataset with (baseline = T) or without (baseline = F) data.

MAVEn dataset with or without baseline data.
}
\description{
\code{read_maven} imports a MAVEn data with or without the baseline data (i.e. Chamber = 0).
}
\examples{
# import with baseline data
#maven_raw <- read_maven(datadir = "data", maven_datafile = "maven_output.csv", baseline = T)

# import without baseline data
#maven <- read_maven(datadir = "data", maven_datafile = "maven_output.csv", baseline = F)
}
