% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sem}
\alias{sem}
\title{Standard error of the mean}
\usage{
sem(x, n)
}
\arguments{
\item{x}{list of values}

\item{n}{number of cycles}
}
\value{
Standard Error of Meatn
}
\description{
Standard error of the mean
}
