% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_activity}
\alias{summarize_activity}
\title{Summarize animal activity data.}
\usage{
summarize_activity(animal_activity, type = "", activity_threshold = "0.5")
}
\arguments{
\item{animal_activity}{Animal activity dataframe extracted from the MAVEn without
baseline using \code{extract_activity}.}

\item{type}{Summarize data by Chamber (="by_chamber") or by Cycle (="by_cycle").}

\item{activity_threshold}{Threshold value to establish activity state.}
}
\value{

}
\description{
Summarize animal activity data.
}
\examples{
#summarize_activity(animal_activity, type = "by_cycle", activity_threshold = 0.5)
#summarize_activity(animal_activity, type = "by_chamber", activity_threshold = 0.5)
}
