% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_metabolism}
\alias{summarize_metabolism}
\title{Summarize metabolism data.}
\usage{
summarize_metabolism(animal_metabolism, type = "")
}
\arguments{
\item{animal_metabolism}{Animal metabolism dataframe extracted from the MAVEn
without baseline using \code{extract_metabolism}.}

\item{type}{Summarize data by Chamber (="by_chamber") or by Cycle (="by_cycle").}
}
\value{
Table of values with calculated median metabolism by cycle or chamber.
}
\description{
Summarize metabolism data.
}
\examples{
#summarize_metabolism(animal_metabolism, type = "by_cycle")
#summarize_metabolism(animal_metabolism, type = "by_chamber")
}
