% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_maven_overview}
\alias{plot_maven_overview}
\title{Plot MAVEn experimental overview data.}
\usage{
plot_maven_overview(maven_raw, maven_experiment = "", outdir = NULL,
  out_filename = "ExpOverview", out_filetype = ".png")
}
\arguments{
\item{maven_raw}{Complete MAVEn dataset including baseline readings.}

\item{maven_experiment}{Experiment name. Applied to both figures and files.}

\item{outdir}{Folder where figure should be stored.}

\item{out_filename}{Figure name.}

\item{out_filetype}{Figure file type.}
}
\description{
\code{plot_maven_overview} generates four plots to track the instrumental
timecourse and requires the full MAVEn dataset (readings + baseline)
generated by setting `baseline` = T in \code{read_maven}. The graphic can be
automatically saved.
}
