% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{upper.ci}
\alias{upper.ci}
\title{Upper confidence interval}
\usage{
upper.ci(x, n)
}
\arguments{
\item{x}{list of values}

\item{n}{number of cycles}
}
\value{

}
\description{
Upper confidence interval
}
